+++ src/main/java/net/atos/projetFinal/repo/AdresseRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import net.atos.projetFinal.model.Adresse;
 
-public interface AdresseRepository extends JpaRepository<Adresse, Integer>{
+public interface AdresseRepository extends JpaRepository<Adresse, Long>{
