+++ src/main/java/net/atos/projetFinal/repo/StatusRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import net.atos.projetFinal.model.Status;
 
-public interface StatusRepository extends JpaRepository<Status, Integer>{
+public interface StatusRepository extends JpaRepository<Status, Long>{
 
 	
 }
